Updated to print leading zeros

+++ include/arbint.h
@@ -2,11 +2,12 @@
 #define ARBINT_H
 
 #include <stdint.h>
+#include <stddef.h>
 #include <stdbool.h>
 
 typedef struct {
     uint64_t *set;
-    unsigned int size;
+    size_t size;
     bool sign;
 } ArbInt;
